// design/source/mebel-next-en/libs/enums/report.enum.ts
export enum ReportGroup {
	MEMBER = 'MEMBER',
	PRODUCT = 'PRODUCT',
	ARTICLE = 'ARTICLE',
}

export enum ReportReason {
	WRONG_DESCRIPTION = 'WRONG_DESCRIPTION',
	POOR_QUALITY = 'POOR_QUALITY',
	FAKE_PRODUCT = 'FAKE_PRODUCT',
	RUDE_AGENT = 'RUDE_AGENT',
	NOT_DELIVERED = 'NOT_DELIVERED',
	OTHER = 'OTHER',
}

export enum ReportStatus {
	PENDING = 'PENDING',
	RESOLVED = 'RESOLVED',
	REJECTED = 'REJECTED',
}